Added crm_history.F90 Module
This change involves major changes to crm_physics.F90. The primary goal
was to remove all addfld and outfld calls that were scattered throughout
crm_physics.F90, but this was not as simple as moving the calls because
of how they were interspersed with other things happening with the CRM
output. So I did some reording of calls after the CRM call, but tests
confirmed that the model is still BFB. I also include a lot of
formatting changes in this commit because the inconsistencies have
driven me crazy for a long time.

	new file:   components/cam/src/physics/crm/crm_history.F90
	modified:   components/cam/src/physics/crm/crm_physics.F90